remove utils from packaging
