+++ builder/testing/src/test/java/io/helidon/builder/testing/Peel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.builder.testing.utils;
+package io.helidon.builder.testing;
 
 import io.helidon.builder.Builder;
